requirements.txt
beautifulsoup4==4.3.2
PyYAML==3.11
requests==2.7.0
slackclient==0.15

